fix: convert jest.config to mjs for NextJS
